removed duped libs
