+++ banner/src/interfaces/Schemas.ts
